company/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-autoslug==1.9.9
django-tinymce==3.7.1
pillow==10.2.0
python-dotenv==1.0.1
sqlparse==0.4.4
